+++ tests/integration/backend_test.py
@@ -56,7 +56,7 @@ from pytket.extensions.quantinuum.backends.api_wrappers import (
     QuantinuumAPI,
     QuantinuumAPIError,
 )
-from pytket.extensions.quantinuum.backends.quantinuum import _ALL_GATES, GetResultFailed
+from pytket.extensions.quantinuum.backends.quantinuum import _ALL_GATES
 from pytket.predicates import CompilationUnit
 from pytket.wasm import WasmFileHandler
 
@@ -530,11 +530,11 @@ def test_leakage_detection(
 
 
 @given(
-    n_shots=st.integers(min_value=1, max_value=10),  # type: ignore
+    n_shots=st.integers(min_value=1, max_value=10),
     n_bits=st.integers(min_value=0, max_value=10),
 )
 @pytest.mark.timeout(120)
-def test_shots_bits_edgecases(n_shots, n_bits) -> None:
+def test_shots_bits_edgecases(n_shots: int, n_bits: int) -> None:
     quantinuum_backend = QuantinuumBackend("H1-1SC", machine_debug=True)
     c = Circuit(n_bits, n_bits)
 
@@ -597,18 +597,17 @@ def test_simulator(
     assert sum(stab_counts.values()) == n_shots
     assert len(stab_counts) == 2
 
-    # test non-clifford circuit fails on stabilizer backend
-    # unfortunately the job is accepted, then fails, so have to check get_result
+    # test non-clifford circuit fails predicate check
+    # when run on stabilizer backend
     non_stab_circ = (
         Circuit(2, name="non_stab_circ").H(0).Rx(0.1, 0).CX(0, 1).measure_all()
     )
     non_stab_circ = stabilizer_backend.get_compiled_circuit(non_stab_circ)
-    broken_handle = stabilizer_backend.process_circuit(
-        non_stab_circ, n_shots, language=language
-    )
 
-    with pytest.raises(GetResultFailed) as _:
-        _ = stabilizer_backend.get_result(broken_handle)
+    with pytest.raises(CircuitNotValidError):
+        _ = stabilizer_backend.process_circuit(
+            non_stab_circ, n_shots, noisy_simulation=False, language=language
+        )
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)